+++ packages/bridge-controller/src/utils/swaps.ts
@@ -1,9 +1,16 @@
 import type { Hex } from '@metamask/utils';
 
+import { formatChainIdToDec } from './caip-formatters';
+import { CHAIN_IDS } from '../constants/chains';
 import {
   ALLOWED_CONTRACT_ADDRESSES,
   SWAPS_CONTRACT_ADDRESSES,
 } from '../constants/swaps';
+import {
+  DEFAULT_TOKEN_ADDRESS,
+  SWAPS_CHAINID_DEFAULT_TOKEN_MAP,
+} from '../constants/tokens';
+import type { FetchFunction } from '../types';
 
 /**
  * Checks if the given contract address is valid for the given chain ID.
@@ -33,4 +40,69 @@ export function isValidSwapsContractAddress(
  */
 export function getSwapsContractAddress(chainId: Hex): string {
   return SWAPS_CONTRACT_ADDRESSES[chainId];
+}
+
+/**
+ * Gets the client ID header.
+ *
+ * @param clientId - The client ID.
+ * @returns The client ID header.
+ */
+function getClientIdHeader(clientId?: string) {
+  if (!clientId) {
+    return undefined;
+  }
+  return {
+    'X-Client-Id': clientId,
+  };
+}
+
+export const API_BASE_URL = 'https://swap.api.cx.metamask.io';
+export const DEV_BASE_URL = 'https://swap.dev-api.cx.metamask.io';
+
+export type SwapsToken = {
+  address: string;
+  symbol: string;
+  name?: string;
+  decimals: number;
+  iconUrl?: string;
+  occurrences?: number;
+};
+
+/**
+ * Fetches token metadata from API URL.
+ *
+ * @param chainId - Current chainId.
+ * @param fetchFn - Fetch function.
+ * @param clientId - Client id.
+ * @returns Promise resolving to an object containing token metadata.
+ */
+export async function fetchTokens(
+  chainId: Hex,
+  fetchFn: FetchFunction,
+  clientId?: string,
+): Promise<SwapsToken[]> {
+  const [apiChainId, apiBaseUrl] =
+    chainId === CHAIN_IDS.LOCALHOST
+      ? [CHAIN_IDS.MAINNET, DEV_BASE_URL]
+      : [chainId, API_BASE_URL];
+
+  const apiDecimalChainId = formatChainIdToDec(apiChainId);
+  const tokenUrl = `${apiBaseUrl}/networks/${apiDecimalChainId}/tokens`;
+
+  const tokens: SwapsToken[] = await fetchFn(tokenUrl, {
+    headers: getClientIdHeader(clientId),
+  });
+
+  const filteredTokens = tokens.filter((token) => {
+    return token.address !== DEFAULT_TOKEN_ADDRESS;
+  });
+
+  const nativeSwapsToken =
+    SWAPS_CHAINID_DEFAULT_TOKEN_MAP[
+      chainId as keyof typeof SWAPS_CHAINID_DEFAULT_TOKEN_MAP
+    ];
+
+  filteredTokens.push(nativeSwapsToken);
+  return filteredTokens;
 }